requirement.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.3
